+++ components/CircleButton/index.module.css
@@ -12,6 +12,4 @@
   border-radius: 50%;
   width: 100px;
   height: 100px;
-  left: 215px;
-  top: 130px;
 }
