+++ components/LandingPageButton/index.module.css
@@ -6,15 +6,15 @@
     font-style: normal;
     font-size: x-large;
     text-align: center;
-    padding: 10px;
-  
+    padding: 1rem;
     border: none;
     color: var(--neutral);
     border-radius: 10px;
-    width: 180px;
-    height: 50px;
-    left: 98px;
-    top: 571px;
+    height: 4rem;
+    width:60%;
+    left: 20%;
+    right:20%;
+    top: 70vh;
     box-shadow: inset 0px 4px 4px rgba(0, 0, 0, 0.25);
   }
