--- components/SignInput/index.module.css
@@ -1,5 +1,0 @@
-.input {
-  position: absolute;
-  left: 40px;
-  top: 180px;
-}
