+++ components/SocialMediaLogo/index.module.css
@@ -1,5 +1,1 @@
-.logo {
-  position: absolute;
-  left: 75px;
-  top: 320px;
-}
+
