+++ container/SignInContainer/index.js
@@ -1,30 +1,43 @@
 import classes from "./index.module.css";
-import SignInput from "../../components/SignInput";
+import SignInput from "../SignInput";
+import { useRouter } from "next/router";
 import SocialMediaLogo from "../../components/SocialMediaLogo";
-import NavigateNextIcon from '@mui/icons-material/NavigateNext';
+import NavigateNextIcon from "@mui/icons-material/NavigateNext";
+import ForgetPassword from "../../pages/ForgetPassword";
 export default function SignInContainer() {
-    return(
-        <div className={classes.block}>
-            <SignInput label= "Email Address" type="email"/>
-            <SocialMediaLogo source="/SignIn/googleIcon.svg"/>
-            <SocialMediaLogo source="/SignIn/FacebookIcon.svg" style={{left:"170px"}}/>
-            <SocialMediaLogo source="/SignIn/appleIcon.svg" style={{left:"265px"}}/>
-            <div style={{position:"absolute",top:400, left:45, color:"#a6a0a9"}}>
-                Forget Password?
-            </div>
-            <div style={{position:"absolute",top:420, left:45, color:"#a6a0a9"}}>
-                New to SE Community?
-            </div>
-            <div style={{
-                position:"absolute",
-                top:450, 
-                left:45, 
-                color:"black",
-                fontSize:"large",
-                fontWeight:700
-                }}>
-                Sign Up <NavigateNextIcon sx={{fontSize:"large"}}/>
-            </div>
+  const router = useRouter();
+  const signUpHandler = () => {
+    router.push("/SignUp");
+  };
+
+  const forgetPasswordHandler = () => {
+    router.push("/ForgetPassword");
+  };
+
+  return (
+    <div className={classes.block}>
+      <div className={classes.input}>
+        <SignInput label="Email Address" type="email" />
+
+        <div className={classes.logos}>
+          <SocialMediaLogo source="/SignIn/googleIcon.svg" />
+          <SocialMediaLogo source="/SignIn/FacebookIcon.svg" />
+          <SocialMediaLogo source="/SignIn/appleIcon.svg" />
         </div>
-    );
+
+        <div className={classes.link}>
+          <div
+            onClick={forgetPasswordHandler}
+          >
+            Forget Password?
+          </div>
+          <div
+            onClick={signUpHandler}
+          >
+            New to SE Community?
+          </div>
+        </div>
+      </div>
+    </div>
+  );
 }
