+++ pages/ForgetPassword/index.js
@@ -1,0 +1,7 @@
+export default function ForgetPassword() {
+    return(
+        <div>
+            Forget password page
+        </div>
+    );
+}
